pyproject.toml
[project]
name = "test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.128.0",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.10.1",
    "python-jose>=3.5.0",
    "python-keycloak>=7.0.0",
    "sqlalchemy>=2.0.45",
]
